Move ProductNotFoundException handling when treated as NOT_FOUND to GlobalExceptionHandler
