--- src/main/java/com/github/jbence1994/webshop/image/ProductPhotoControllerExceptionHandler.java
@@ -1,18 +1,0 @@
-package com.github.jbence1994.webshop.image;
-
-import com.github.jbence1994.webshop.common.ErrorDto;
-import com.github.jbence1994.webshop.product.ProductNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice(basePackageClasses = ProductPhotoController.class)
-public class ProductPhotoControllerExceptionHandler {
-
-    // FIXME: Duplicate... Move to common.
-    @ExceptionHandler(exception = ProductNotFoundException.class)
-    public ResponseEntity<ErrorDto> handleProductNotFoundException(ProductNotFoundException exception) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorDto(exception.getMessage()));
-    }
-}
