--- src/main/java/com/github/jbence1994/webshop/product/ProductControllerExceptionHandler.java
@@ -1,16 +1,0 @@
-package com.github.jbence1994.webshop.product;
-
-import com.github.jbence1994.webshop.common.ErrorDto;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice(basePackageClasses = ProductController.class)
-public class ProductControllerExceptionHandler {
-
-    @ExceptionHandler(exception = ProductNotFoundException.class)
-    public ResponseEntity<ErrorDto> handleProductNotFoundException(ProductNotFoundException exception) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorDto(exception.getMessage()));
-    }
-}
