+++ src/test/java/com/github/jbence1994/webshop/common/GlobalExceptionHandlerTests.java
@@ -3,6 +3,7 @@ package com.github.jbence1994.webshop.common;
 import com.github.jbence1994.webshop.image.ImageDeletionException;
 import com.github.jbence1994.webshop.image.ImageUploadException;
 import com.github.jbence1994.webshop.image.InvalidFileExtensionException;
+import com.github.jbence1994.webshop.product.ProductNotFoundException;
 import com.github.jbence1994.webshop.user.UserNotFoundException;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
@@ -83,6 +84,15 @@ public class GlobalExceptionHandlerTests {
         assertThat(result.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
         assertThat(result.getBody(), not(nullValue()));
         assertThat(result.getBody().error(), equalTo("The file must not be empty."));
+    }
+
+    @Test
+    public void handleProductNotFoundExceptionTest() {
+        var result = globalExceptionHandler.handleProductNotFoundException(new ProductNotFoundException(1L));
+
+        assertThat(result.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+        assertThat(result.getBody(), not(nullValue()));
+        assertThat(result.getBody().error(), equalTo("No product was found with the given ID: #1."));
     }
 
     @Test
